Core: fix Token.h enum issue with CLang

+++ Source/Core/Container/Token.h
@@ -90,7 +90,7 @@ public:
     typedef _Allocator allocator_type;
     typedef TokenFactory<Token> factory_type;
 
-    enum { Sensitiveness = _CaseSensitive };
+    enum { Sensitiveness = size_t(_CaseSensitive) };
 
     Token();
     ~Token();
